requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.2
python-dotenv==1.0.1
requests==2.32.4
rich==14.3.3
urllib3==2.2.3
